removed unnecessary checks
Assuming that the email is not in the database, there is no need to check for getStatus().
It would be easier to check if the object is null.
Also there is no need for getStatus() (yet) since the only option is "register" (email not registered -> emailEntity is null) or "login" (email registered already ->  emailEntity wont be null)