Create EmailRepository.java
This interface connects to your mongodb and finds/updates etc. documents in your database.
Repositories are magic in Spring Boot. They will automatically know what to return (from the return type), what to search for (from the method signature) and with what value to search for (from the parameter). It will automatically search the database, put the found user (if any user is found) in a new EmailEntity object and return it. If nothing is found it will return null