Invite Token Validation
Every JWT token (eyXXXXXX -> see accessToken) is given a subject and expiration date.
By the subject we can see what type of JWT token it is.
If the subject is a plain username it is an accessToken.
If the subject is INVITE-<username> we know its a company invitation.
We will strap the username from the accessToken from the user who wants to accept the invitation and pass it to our validateInviteToken to check if the user matches.
If everything is okay we only need to check if the invitation has expired or not